Extract Reindeer class in an external file to be used in part 2 of the day 14.
