+++ 2015/Day14/src/main.cpp
@@ -1,60 +1,8 @@
-#include "string_manipulation.hpp"
+#include "Reindeer.hpp"
 #include <cassert>
 #include <fstream>
 #include <iostream>
 #include <algorithm>
-#include <sstream>
-
-using Speed = size_t;
-using Position = size_t;
-using Time = size_t;
-
-class Reindeer
-{
-public:
-    Reindeer (const std::string& name_, const Speed flyingSpeed_, const Time flyingTime_, const Time restTime_) : name(name_), flyingSpeed(flyingSpeed_), flyingTime(flyingTime_), restTime(restTime_), timeBeforeChangingState(flyingTime_) { }
-    ~Reindeer () = default;
-
-    void oneSecondPassed ()
-    {
-        if(timeBeforeChangingState == 0)
-        {
-            isFlying = !isFlying;
-            timeBeforeChangingState = isFlying ? flyingTime : restTime;
-        }
-
-        if(isFlying)
-        {
-            position += flyingSpeed;
-        }
-        --timeBeforeChangingState;
-    }
-
-    Position getCurrentPosition () const { return position; }
-
-private:
-    std::string name;
-    Speed flyingSpeed{0};
-    Time flyingTime{0};
-    Time restTime{0};
-
-    Position position{0};
-    Time currentTimePassed{0};
-    Time timeBeforeChangingState{0};
-    bool isFlying{true};
-};
-
-Reindeer extractInformationFrom (const std::string& line)
-{
-    std::istringstream stream(line);
-    std::string reindeerName, can, fly, km_s, for_, seconds, but, then, must, rest, for__, seconds_;
-    Speed flyingSpeed;
-    Time flyingTime;
-    Time restTime;
- 
-    stream >> reindeerName >> can >> fly >> flyingSpeed >> km_s >> for_ >> flyingTime >> seconds >> but >> then >> must >> rest >> for__ >> restTime >> seconds_;
-    return Reindeer(reindeerName, flyingSpeed, flyingTime, restTime);
-}
 
 int main (int argc, char** argv)
 {
